Updates to match design
This starts focusing in on the design and also the lazy loading of the payment providers so it's not so jarring when they show up.

This did uncover a race condition where we overwriting the callback when multiple things tried to load the braintree scripts. It got fixed in the lazy loader service.