Replace handler Promises with PromisedSingletons
For the braintree libraries, we only want to load them once and we load them as Promises as-needed.

This presents a problem when several consumers request the clients at once. The Promises don't automatically chain so you have to do some gatekeeping to make sure only one instance gets created.

This adds a class called `PromisedSingleton` that has 2 methods: `get()` and `reset()`. It gets initialized with a Promise that generates the singleton and when `get()` is first called, it executes the Promise, caches its results and returns it to the caller.

If more callers call it in the interim, it chains the promises.

This reduces so much boilerplate!