Update Error Messaging (#2)
This updates the error messaging and error modal to be more useful.

Previously we were displaying an error message that looked like An error occurred: , because we weren't getting the proper error messaging from Braintree. This adds a "Questions?" link at the bottom of the error modal and makes the text more friendly.