Fix Recaptcha Race Condition (#4)
Previously we were relying on the Recaptcha library to be loaded by the host, but this was causing race conditions if the Donation Form got initialized before Recaptcha loaded. It caused the Donate button to stop working but threw no errors.

Now we lazy load the Recaptcha client like we do the other clients, but add an onload callback to the library so we know when it's been loaded.