Reset captcha instead of failing if captcha already in progress
If the user dismisses the captcha, it's still "in progress" and we don't get notified thet the dismissal has occurred so we have no way to reset it ahead of time.

Now if the user dismisses it, then starts a new one, we just reset it right there so it doesn't fail.