Update the Contact Form and Validation Behavior (#21)
This changes the contact form to only collect email, name, and zip code and adds little red asterisks to signify they're all required.

It also changes the validation behavior. Previously we left the Donate button disabled until all of the fields were valid. Now we leave the Donate button active and show the user errors when they click it.